finish form's layout
